+++ src/library/pkgbuild/R/find-package-root.R
@@ -18,8 +18,10 @@ find_package_root <- function(path = ".") {
   max_depth <- 100
   for (i in 1:max_depth) {
     dsc_path <- file.path(cur_path, "DESCRIPTION")
-    if (file.exists(dsc_path) &&
-      any(grepl("^Package: ", readLines(dsc_path)))) {
+    if (
+      file.exists(dsc_path) &&
+        any(grepl("^Package: ", readLines(dsc_path)))
+    ) {
       return(cur_path)
     } else if (is_root(cur_path)) {
       stop(errmsg)