+++ src/library/pkgbuild/R/rtools-registry.R
@@ -9,9 +9,24 @@ scan_registry_for_rtools <- function(debug = FALSE) {
   if (debug) cat("Scanning registry...\n")
 
   keys <- c(
-    read_registry("SOFTWARE\\R-core\\Rtools", hive = "HCU", view = "32-bit", maxdepth = 2),
-    read_registry("SOFTWARE\\R-core\\Rtools", hive = "HLM", view = "32-bit", maxdepth = 2),
-    read_registry("SOFTWARE\\R-core\\Rtools", hive = "HLM", view = "64-bit", maxdepth = 2)
+    read_registry(
+      "SOFTWARE\\R-core\\Rtools",
+      hive = "HCU",
+      view = "32-bit",
+      maxdepth = 2
+    ),
+    read_registry(
+      "SOFTWARE\\R-core\\Rtools",
+      hive = "HLM",
+      view = "32-bit",
+      maxdepth = 2
+    ),
+    read_registry(
+      "SOFTWARE\\R-core\\Rtools",
+      hive = "HLM",
+      view = "64-bit",
+      maxdepth = 2
+    )
   )
 
   if (is.null(keys)) {
@@ -24,8 +39,10 @@ scan_registry_for_rtools <- function(debug = FALSE) {
     version <- names(keys)[[i]]
     key <- keys[[version]]
     if (!is.list(key) || is.null(key$InstallPath)) next
-    install_path <- normalizePath(key$InstallPath,
-      mustWork = FALSE, winslash = "/"
+    install_path <- normalizePath(
+      key$InstallPath,
+      mustWork = FALSE,
+      winslash = "/"
     )
 
     if (debug) cat("Found", install_path, "for", version, "\n")